+++ src/blocks/no-results/no-results.css
@@ -3,7 +3,7 @@
 @import url('./__image/no-results__image.css');
 
 .no-results {  
-    width: auto;
+    width: auto;    
     padding-top: 86px;
     padding-bottom: 80px;
     display: flex;
