+++ src/blocks/page/page.css
@@ -1,6 +1,6 @@
 .page {
     margin: 0;
-    width: 100%;  
+    width: 100%;     
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
 }
