+++ src/blocks/preloader/preloader.css
@@ -5,7 +5,7 @@
 	display: none;
 	position: relative;
 	background: #F5F6F7;
-	height: 282px;
+	height: 282px;	
 }
 
 @keyframes spin {
