+++ src/scripts/cardList.js
@@ -1,18 +1,17 @@
+import * as url from "../images/news.jpg"
+
 export class CardList {
     constructor(container, cardsArray) {
-        this.container = container;
-        this.cards = cardsArray;   
-        this.renderedCards = 0;         
-        
-        this.clearList();
-        this.render();    
+        this._container = container;
+        this._cards = cardsArray;   
+        this._renderedCards = 0;         
     }
 
     _createElement(type, userClasses) {
         const el = document.createElement(type);
         el.classList.add(userClasses);
         return el;
-    }
+    } 
 
     _transformDate(date) {        
         const months = {
@@ -38,16 +37,20 @@ export class CardList {
         return transformedDate;
     }
     
-    _makeCard(cardsItem) {           
+    _makeCard(cardsItem) {          
         const cardLink = this._createElement('a', 'link', 'results__card-link');
         cardLink.setAttribute('href', `${cardsItem.url}`);
         cardLink.setAttribute('target', "_blank")
 
         const card = this._createElement('div', 'results__card');
 
-        const cardImage = this._createElement('img', 'results__card-image');        
-        cardImage.setAttribute('src', `${cardsItem.urlToImage}`);
-        cardImage.setAttribute('alt', "Картинка для статьи")
+        const cardImage = this._createElement('img', 'results__card-image');
+        
+        cardImage.setAttribute('src', `${cardsItem.urlToImage}`);           
+        cardImage.onerror = function() {
+            cardImage.src = url;
+        }
+        cardImage.setAttribute('alt', "Картинка для статьи")      
 
         const cardInfo = this._createElement('div', 'results__card-info');
 
@@ -75,30 +78,30 @@ export class CardList {
     } 
 
     clearList() {
-        while (this.container.hasChildNodes()) {
-            this.container.removeChild(this.container.lastChild);
-        }
+        Array.from(this._container.childNodes).forEach((item) => {            
+            this._container.removeChild(item)
+        })
     }
     
     _addCard(cardsItem) { 
-        this.container.appendChild(this._makeCard(cardsItem)) 
+        this._container.appendChild(this._makeCard(cardsItem))         
     }
     
     render() {         
         const cardsToRender = 3;
         const moreButton = document.querySelector('.results__more-button');
         
-        if (this.renderedCards + cardsToRender >= this.cards.length) {
-            for (let i = this.renderedCards; i < this.cards.length; i++) {
-                this._addCard(this.cards[i]);            
+        if (this._renderedCards + cardsToRender >= this._cards.length) {
+            for (let i = this._renderedCards; i < this._cards.length; i++) {
+                this._addCard(this._cards[i]);            
             }    
-            this.renderedCards = this.cards.length;
+            this._renderedCards = this._cards.length;
             moreButton.style.display = 'none';
         } else {
-            for (let i = this.renderedCards; i < this.renderedCards + cardsToRender; i++) {
-                this._addCard(this.cards[i]);            
+            for (let i = this._renderedCards; i < this._renderedCards + cardsToRender; i++) {
+                this._addCard(this._cards[i]);            
             }  
-            this.renderedCards = this.renderedCards + cardsToRender;  
+            this._renderedCards = this._renderedCards + cardsToRender;  
         }         
     }
 }
