--- src/scripts/get_current_date.js
@@ -1,9 +1,0 @@
-export function getCurrentDate() {
-    let date = new Date(); 
-
-    let month = date.getMonth() + 1;
-    let day = date.getDate();  
-    let year = date.getFullYear();
-    
-    return (year + '-' + month + '-' + day)
-}
